Move MAGICC test example into docs
